// src/libs/api/schemas/pedestrianRouteRequest.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */

/**
 * 보행자 경로 안내 요청 정보를 담는 DTO입니다.
 */
export interface PedestrianRouteRequest {
  /** 각도 (0~359) */
  angle?: number;
  /** 목적지 명칭 (UTF-8 URL 인코딩 필수) */
  endName: string;
  /** 목적지 POI ID */
  endPoiId?: string;
  /** 목적지 X 좌표 (경도) */
  endX: number;
  /** 목적지 Y 좌표 (위도) */
  endY: number;
  /** 경유지의 X, Y 좌표 (최대 5개, '_'로 구분) */
  passList?: string;
  /** 요청 좌표계 (WGS84GEO, EPSG3857, KATECH) */
  reqCoordType?: string;
  /** 응답 좌표계 (WGS84GEO, EPSG3857, KATECH) */
  resCoordType?: string;
  /** 경로 탐색 옵션 (0: 추천, 4: 추천+대로 우선, 10: 최단, 30: 최단거리+계단 제외) */
  searchOption?: number;
  /** 정렬 순서 (index, custom) */
  sort?: string;
  /** 진행 속도 (km/h, 0 이상) */
  speed?: number;
  /** 출발지 명칭 (UTF-8 URL 인코딩 필수) */
  startName: string;
  /** 출발지 X 좌표 (경도) */
  startX: number;
  /** 출발지 Y 좌표 (위도) */
  startY: number;
}