// src/libs/api/schemas/poi.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { PoiBizName } from './poiBizName';
import type { PoiCollectionType } from './poiCollectionType';
import type { PoiDataKind } from './poiDataKind';
import type { PoiDesc } from './poiDesc';
import type { PoiDetailAddrName } from './poiDetailAddrName';
import type { PoiDetailBizName } from './poiDetailBizName';
import type { PoiDetailInfoFlag } from './poiDetailInfoFlag';
import type { PoiEvChargers } from './poiEvChargers';
import type { PoiFirstBuildNo } from './poiFirstBuildNo';
import type { PoiFirstNo } from './poiFirstNo';
import type { PoiFrontLat } from './poiFrontLat';
import type { PoiFrontLon } from './poiFrontLon';
import type { PoiGroupSubLists } from './poiGroupSubLists';
import type { PoiId } from './poiId';
import type { PoiLowerAddrName } from './poiLowerAddrName';
import type { PoiLowerBizName } from './poiLowerBizName';
import type { PoiMiddleAddrName } from './poiMiddleAddrName';
import type { PoiMiddleBizName } from './poiMiddleBizName';
import type { PoiMlClass } from './poiMlClass';
import type { PoiName } from './poiName';
import type { PoiNavSeq } from './poiNavSeq';
import type { PoiNewAddressList } from './poiNewAddressList';
import type { PoiNoorLat } from './poiNoorLat';
import type { PoiNoorLon } from './poiNoorLon';
import type { PoiParkFlag } from './poiParkFlag';
import type { PoiPkey } from './poiPkey';
import type { PoiRadius } from './poiRadius';
import type { PoiRoadName } from './poiRoadName';
import type { PoiRpFlag } from './poiRpFlag';
import type { PoiSecondBuildNo } from './poiSecondBuildNo';
import type { PoiSecondNo } from './poiSecondNo';
import type { PoiTelNo } from './poiTelNo';
import type { PoiUpperAddrName } from './poiUpperAddrName';
import type { PoiUpperBizName } from './poiUpperBizName';
import type { PoiZipCode } from './poiZipCode';

export interface Poi {
  /** 대표 업종명 */
  bizName?: PoiBizName;
  /** 문서의 컬렉션(address, poi) 출처 */
  collectionType?: PoiCollectionType;
  /** 데이터 구분자 */
  dataKind?: PoiDataKind;
  /** 검색된 관심 장소(POI)에 대한 소개 */
  desc?: PoiDesc;
  /** 표출 주소 세분류명(동/호/층) */
  detailAddrName?: PoiDetailAddrName;
  /** 업종 상세 분류명 */
  detailBizName?: PoiDetailBizName;
  /** 관심 장소(POI) 상세 정보 유무 */
  detailInfoFlag?: PoiDetailInfoFlag;
  /** 충전기 목록 */
  evChargers?: PoiEvChargers;
  /** 건물번호의 본번(새주소) */
  firstBuildNo?: PoiFirstBuildNo;
  /** 주소 본번 */
  firstNo?: PoiFirstNo;
  /** 장소(시설물) 입구 위도 좌표 */
  frontLat?: PoiFrontLat;
  /** 장소(시설물) 입구 경도 좌표 */
  frontLon?: PoiFrontLon;
  /** 검색한 관심 장소(POI)의 그룹 하위 목록 */
  groupSubLists?: PoiGroupSubLists;
  /** 관심 장소(POI) ID */
  id?: PoiId;
  /** 표출 주소 소분류명(읍/면/동) */
  lowerAddrName?: PoiLowerAddrName;
  /** 업종 소분류명 */
  lowerBizName?: PoiLowerBizName;
  /** 표출 주소 중분류명(시/군/구) */
  middleAddrName?: PoiMiddleAddrName;
  /** 업종 중분류명 */
  middleBizName?: PoiMiddleBizName;
  /** 산/대지 구분(구주소) */
  mlClass?: PoiMlClass;
  /** 장소명(시설물 등) 및 업체명 */
  name?: PoiName;
  /** 장소(시설물)의 입구 일련번호 */
  navSeq?: PoiNavSeq;
  /** 새주소 목록 */
  newAddressList?: PoiNewAddressList;
  /** 중심점 위도 좌표 */
  noorLat?: PoiNoorLat;
  /** 중심점 경도 좌표 */
  noorLon?: PoiNoorLon;
  /** 주차 가능 유무 */
  parkFlag?: PoiParkFlag;
  /** 관심 장소(POI) 식별자 */
  pkey?: PoiPkey;
  /** 요청 좌표에서 떨어진 거리 */
  radius?: PoiRadius;
  /** 도로명 */
  roadName?: PoiRoadName;
  /** 길안내 요청에 필요한 값 */
  rpFlag?: PoiRpFlag;
  /** 건물번호의 부번(새주소) */
  secondBuildNo?: PoiSecondBuildNo;
  /** 주소 부번 */
  secondNo?: PoiSecondNo;
  /** 전화번호 */
  telNo?: PoiTelNo;
  /** 표출 주소 대분류명(시/도) */
  upperAddrName?: PoiUpperAddrName;
  /** 업종 대분류명 */
  upperBizName?: PoiUpperBizName;
  /** 우편번호 */
  zipCode?: PoiZipCode;
}